# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraApi",
        "description": "LaraApi project API documents",
        "contact": {
            "email": "russell57260620@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/resume": {
            "get": {
                "tags": [
                    "Resume"
                ],
                "summary": "index",
                "description": "Display a listing of the resumes table.",
                "operationId": "getResumeIndex",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Resume"
                ],
                "summary": "store",
                "description": "Store a newly created resume in resumes table.",
                "operationId": "postResumeStore",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "head": {
                "tags": [
                    "Resume"
                ],
                "summary": "index",
                "description": "Display a listing of the resumes table.",
                "operationId": "headResumeIndex",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            }
        },
        "/api/resume/{id}": {
            "get": {
                "tags": [
                    "Resume"
                ],
                "summary": "show",
                "description": "Display the specified resume with resume id.",
                "operationId": "getResumeShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Resume ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Resume"
                ],
                "summary": "update",
                "description": "Update the specified resume in resumes table.",
                "operationId": "putResumeUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Resume ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Resume"
                ],
                "summary": "destroy",
                "description": "Remove the specified resume from resumes table.",
                "operationId": "deleteResumeDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Resume ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "head": {
                "tags": [
                    "Resume"
                ],
                "summary": "show",
                "description": "Display the specified resume with resume id.",
                "operationId": "headResumeShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Resume ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Resume"
                ],
                "summary": "update",
                "description": "Update the specified resume in resumes table.",
                "operationId": "patchResumeUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Resume ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    },
...
}